raft: make election and heartbeat ticks of type int64
In the near future, we want to keep incrementing the electionElapsed
until we hear from a leader. In order to avoid any chances of an integer
overflow, this commit changes the electionElapsed (along with other
symmetrical variables from int to int64)

Epic: None

Release note: None

+++ pkg/kv/kvserver/testing_knobs.go
@@ -279,7 +279,7 @@ type StoreTestingKnobs struct {
 	// RefreshReasonTicksPeriod overrides the default period over which
 	// pending commands are refreshed. The period is specified as a multiple
 	// of Raft group ticks.
-	RefreshReasonTicksPeriod int
+	RefreshReasonTicksPeriod int64
 	// DisableProcessRaft disables the process raft loop.
 	DisableProcessRaft func(roachpb.StoreID) bool
 	// DisableLastProcessedCheck disables checking on replica queue last processed times.
